+++ tests/unit_tests/test_schematic.py
@@ -33,9 +33,6 @@ from skidl.errors import PlacementFailure, RoutingFailure
 
 
 sch_options = {}
-# seed = int(time.time())
-# sch_options["seed"] = seed
-# print(f"Random seed = {seed}")
 sch_options["retries"] = 3
 # sch_options["allow_routing_failure"] = True
 # sch_options["pt_to_pt_mult"] = 1  # HACK: Ad-hoc value.